requirements.txt
requests
google-generativeai
aiohttp
python-dotenv